mushroom-mvp/print-service/pyproject.toml
[tool.poetry]
name = "mushroom-print-service"
version = "1.0.0"
description = "Label printing service for Mushroom MVP using NIIMBOT B1"
authors = ["Marcio"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
flask-cors = "^4.0.0"
pillow = "^10.0.0"
qrcode = {extras = ["pil"], version = "^7.4.0"}
bleak = "^0.21.0"

[tool.poetry.scripts]
print-server = "app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
